Ajout des fonctionnalités de gestion des catégories : création, édition, suppression et affichage en back-office, ainsi que la mise à jour des routes associées.

+++ routes/web.php
@@ -1,45 +1,54 @@
 <?php
 
 use Illuminate\Support\Facades\Route;
-
 use App\Http\Controllers\CategoryController;
 use App\Http\Controllers\ArticleController;
+
+/*
+|--------------------------------------------------------------------------
+| ROUTES PUBLIQUES (FRONT-OFFICE)
+|--------------------------------------------------------------------------
+*/
 
 Route::get('/', function () {
     return view('home');
 });
 
+// Articles
 Route::get('/articles', [ArticleController::class, 'index'])->name('articles.index');
+Route::get('/articles/{slug}', [ArticleController::class, 'show'])->name('articles.show');
 
-Route::get('/articles/{slug}',[ArticleController::class,'show'])->name('articles.show');
-
+// Catégories (Vue publique si vous affichez la liste des catégories aux visiteurs)
 Route::get('/categories', [CategoryController::class, 'index'])->name('categories.index');
 
 
-    
-// Route pour afficher le formulaire d'édition
-Route::get('/categories/{category}/edit', [CategoryController::class, 'edit'])->name('categories.edit');
+/*
+|--------------------------------------------------------------------------
+| ROUTES ADMINISTRATION (BACK-OFFICE)
+|--------------------------------------------------------------------------
+*/
 
-// Route pour traiter la suppression
-Route::delete('/categories/{category}', [CategoryController::class, 'destroy'])->name('categories.destroy');
-
-// --- ESPACE ADMINISTRATION ARTICLES ---
-
-// 1. Afficher la liste des articles (Admin)
+// --- ESPACE ADMIN : ARTICLES ---
 Route::get('/admin/articles', [ArticleController::class, 'adminIndex'])->name('admin.articles.index');
-
-// 2. Formulaire de création (Écran 5)
 Route::get('/admin/articles/creer', [ArticleController::class, 'create'])->name('admin.articles.create');
-
-// 3. Action de sauvegarde de la création en BDD
 Route::post('/admin/articles', [ArticleController::class, 'store'])->name('admin.articles.store');
-
-// 4. Formulaire d'édition (Écran 5 pré-rempli)
 Route::get('/admin/articles/{slug}/modifier', [ArticleController::class, 'edit'])->name('admin.articles.edit');
-
-// 5. Action de mise à jour en BDD (après édition)
 Route::put('/admin/articles/{slug}', [ArticleController::class, 'update'])->name('admin.articles.update');
-
-// 6. Action de suppression de l'article (Écran 6)
 Route::delete('/admin/articles/{id}', [ArticleController::class, 'destroy'])->name('admin.articles.destroy');
 
+
+// --- ESPACE ADMIN : CATÉGORIES ---
+
+// 1. Liste des catégories en admin
+Route::get('/admin/categories', [CategoryController::class, 'adminIndex'])->name('admin.categories.index');
+
+// 2. Formulaire de création + sauvegarde
+Route::get('/admin/categories/creer', [CategoryController::class, 'create'])->name('admin.categories.create');
+Route::post('/admin/categories', [CategoryController::class, 'store'])->name('admin.categories.store');
+
+// 3. Formulaire de modification + sauvegarde
+Route::get('/admin/categories/{category}/modifier', [CategoryController::class, 'edit'])->name('admin.categories.edit');
+Route::put('/admin/categories/{category}', [CategoryController::class, 'update'])->name('admin.categories.update');
+
+// 4. Suppression
+Route::delete('/admin/categories/{category}', [CategoryController::class, 'destroy'])->name('admin.categories.destroy');
